Rename module: 'config' -> 'configs'
